# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Spool API",
    "version": "1.0.0",
    "description": "The REST API of a Spool server: a self-hosted git forge that keeps every repository in object storage. Git smart-HTTP endpoints (clone/fetch/push at /{org}/{repo}.git) speak the standard git wire protocol and are not described here. Every repository is private to the organization that owns it: a request with no credential answers 401 on every repository, changeset, search and profile route, and a credential that may not read a repository gets the same 404 a repository that does not exist gets. The exceptions are the sign-in routes under /v1/auth, the health and metrics probes, the inbound webhook receiver, the runner routes (which take their own credentials) and the signed CDN pack URL. Errors return JSON {\"error\": \"…\"} with the matching status. Optimistic-concurrency conflicts return 409 with the current tip."
  },
  "servers": [
    {
      "url": "https://spool.example.com",
      "description": "Your server; set by STRATUM_PUBLIC_URL."
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token of the form weft_<id>_<secret>. Scopes: org:admin, org:read, repo:read, repo:write. A token acts only in the organization it was minted in: presented anywhere else it has no access, and a repository there answers 404. Repo-bound tokens are restricted to one repository and cannot search."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "stratum_session",
        "description": "Browser session from POST /v1/auth/login. An alternative to a bearer token for the dashboard; a bearer token wins when both are presented. A session spans every organization its person belongs to, which is why anything that crosses two namespaces — forking, opening a change from a fork — takes one. Routes under /v1/auth use this scheme."
      }
    },
...
  },
  "paths": {
...
    "/v1/orgs/{org}/repos/{repo}/changes": {
      "post": {
        "summary": "Register a branch tip as a change (one commit = one review unit, keyed by its Change-Id trailer). repo:read when proposing from a fork via `source`; repo:write when proposing from the repository itself.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "from"
                ],
                "properties": {
                  "from": {
                    "type": "string",
                    "description": "Rev (usually a branch) whose tip becomes the patchset"
                  },
                  "target": {
                    "type": "string",
                    "description": "Branch the change intends to land on; the repo default if absent"
                  },
                  "source": {
                    "type": "string",
                    "description": "owner/name of a fork the commits live in, when they do not live in the repository being targeted. This is how somebody with no push credential contributes: fork, push to the repository you own, open the change here. With source set the caller needs only repo:read on the target; without it they still need repo:write, because those commits could only be in the target if an authorised push put them there. The source must be readable by the caller and must actually be a fork of this repository."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "New change or patchset: { change, patchset }"
          },
          "200": {
            "description": "Idempotent ack of a commit already registered"
          },
          "409": {
            "description": "The change is landing/landed/abandoned and refuses new patchsets"
          },
          "403": {
            "description": "Proposing from this repository without write access; the message names forking as the way through"
          },
          "404": {
            "description": "The target, or a named source, is masked from this caller"
          }
        }
      },
      "get": {
        "summary": "List changes, newest first",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "Whitespace-separated name:value filters. `is:open` / `is:landing` / `is:landed` / `is:abandoned`; `author:@me` or `author:<email>`; `needs:my-approval`. `repo:<name>` is understood only on the org-wide list and is refused here. An unrecognised term is a 400 that names it — nothing is ever silently ignored, because a filter that quietly does nothing is how somebody concludes there is no work waiting for them. A well-shaped address that names no account here is an empty page rather than a refusal, so this list is not an address oracle. `author:@me` and `needs:my-approval` need a person: anonymous is 401 and a service token is 403.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "description": "The same filter as q=is:…. Given both, they must agree, or 400.",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "landing",
                "landed",
                "abandoned"
              ]
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "description": "Keyset cursor: the `next` of the previous page, opaque. Not an offset — changes arrive at the top of this ordering, so an offset would skip a row that shifted down into it. A value this list did not mint is a 400.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Clamped to 1..500, and to 50 when the query contains needs:my-approval, which costs an OWNERS resolution per row examined.",
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ changes: [{key, title, state, land_verdict, patchset, …}], next } — `next` is the cursor for the following page, or null at the end. With needs:my-approval the page may be shorter than `limit`, or empty, with `next` still set: the cursor is where the query stopped, not where the surviving rows did, so a client keeps walking."
          },
          "400": {
            "description": "An unrecognised or repeated q term (the message names it), a state and an is: that disagree, or an `after` this list did not mint"
          },
          "401": {
            "description": "author:@me or needs:my-approval with nobody signed in"
          },
          "403": {
            "description": "author:@me or needs:my-approval under a service token, which is nobody's reviewer"
          }
        }
      }
    },
...
    "/v1/orgs/{org}/changes": {
      "get": {
        "summary": "Every change in the organization the caller may read, newest first — the org-wide list the changeset picker reads, so composing does not begin with a request per repository. Needs a credential: with none the answer is 401. No org-level scope is required beyond that: each repository is authorized exactly as its own changes list is, so a token bound to one repository sees that repository, and a credential from another organization is answered 404, exactly as a name that is nobody's organization is.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "The same grammar as the per-repo list, plus `repo:<name>`, which narrows to one repository. A name this caller may not read is an empty page, never a refusal that would confirm the repository exists. An unrecognised term is a 400 that names it.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "description": "The same filter as q=is:…. Given both, they must agree, or 400.",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "landing",
                "landed",
                "abandoned"
              ]
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "description": "Keyset cursor: the `next` of the previous page, opaque. A value this list did not mint is a 400.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Clamped to 1..500, and to 50 when the query contains needs:my-approval, which costs an OWNERS resolution per row examined.",
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ changes: [{key, title, state, patchset, repo, changeset, …}], next } — `repo` names the repository the change is in, and `changeset` is the key of the changeset already holding it, or null when it is free to be composed; `viewer_write` says whether the caller may push to `repo` — the repository row's own answer, so a picker can grey out a change the caller could not compose or land. `next` is the keyset cursor for the following page, or null at the end"
          },
          "400": {
            "description": "Unknown state, an unrecognised or repeated q term (the message names it), a state and an is: that disagree, or an `after` this list did not mint"
          },
          "401": {
            "description": "no credential, or one that does not authenticate — for an organization that exists and one that does not alike"
          },
          "403": {
            "description": "author:@me or needs:my-approval under a service token, which is nobody's reviewer"
          },
          "404": {
            "description": "no such organization, or the caller is not in it — the two are not told apart"
          }
        }
      }
    },
...
    "/v1/orgs/{org}/changesets/{changeset}": {
      "get": {
        "summary": "One changeset with its members, edges and landing order. repo:read on every member; otherwise 404, whether or not it exists.",
        "responses": {
          "200": {
            "description": "{ key, title, body, state, created_at, updated_at, viewer_write, reverts, reverted_by: [key], members: [{ repo, change }], edges: [{ from, to }], order: [{ repo, change }], landing, composition, checks: [{ repo, name, state, detail_url, run }] } — `viewer_write` is whether the caller may land, revert, abandon or edit the set: write on every member, which is what each of those routes needs, so a client can withhold controls that would be refused with a masked 404. `order` is the landing order the edges imply, ties by the order members were added. `landing` is null until the changeset has been asked to land, then { id, attempt, started_at, finished_at, outcome, members: [{ repo, change, ref, old, new, state, note }] }: the plan in landing order with each member's progress — `state` is `pending`, `done`, `failed` or `reverted`; `note` says why a step failed, names the revert commit of a reverted one, or says why a landed member could not be reverted; `outcome` is null while landing, then `landed` or `failed`; `attempt` counts the jobs that have driven it. `reverts` is the key of the changeset this one was made to revert (`POST …/revert`), or null; `reverted_by` lists the keys of the changesets made to revert this one, oldest first, whatever state they are in. `composition` identifies this set of members at these commits — a hash over each member's repo and its latest patchset commit — over the members listed, so a member whose repository has been deleted is left out of it as it is of `members` — and is null when no member is left or one has no patchset yet; a new patchset on any member, or a membership change, is a new composition, and its runs supersede the previous one's. `checks` is the composed CI for that current composition only, ordered by repo then name: one row per job of a workflow whose `on:` contains `changeset`, `name` being `<workflow> / <job>` as on a commit, `state` one of queued, running, passing, failing, cancelled, skipped, and `run` the workflow run it came from. Composed verdicts are never written to the member commit's `check_runs` — a file with `on: [change, changeset]` would otherwise report the composed answer as the member's own."
          },
          "404": {
            "description": "No such changeset, or a member the caller may not read"
          }
        }
      }
    },
...
    "/v1/orgs/{org}/changesets/{changeset}/revert": {
      "post": {
        "summary": "Revert a landed changeset: a new changeset, one call. repo:write on every member.",
        "description": "Body { key, title?, body? }. In every repository a member of this changeset landed in, makes one commit on a new `revert/{key}` branch off the target branch **as it is now** that puts back every path the member changed — the deleted file returns, the rewritten one is its old self at its old mode, the added ones are gone — and leaves every other path alone, so work landed since is kept. Each commit is registered as a change on the member's target exactly as a push would be (pinned, CI triggered, owners notified), and the changes are composed into changeset `key` with this changeset's edges **reversed**: what landed after its dependency is undone before it. Nothing about this changeset moves, and no trunk moves: the new changeset is reviewed under the same OWNERS and landed with `POST …/land` like any other. Only members recorded as landed are reverted — every member of a `landed` changeset, and of a `failed` one the members the unwind left landed (`GET …/changesets/{changeset}` shows them `done` with a `not reverted` note). Every member is checked before any branch is written, and any one refusal refuses the whole call: a path somebody has changed on trunk since the landing, a target branch that is gone, a `revert/{key}` branch already there, a member whose repository has been deleted. Title defaults to `Revert \"{title}\"`, body to `Reverts changeset {key}.`.",
        "responses": {
          "201": {
            "description": "The new changeset, as `GET …/changesets/{key}` returns it, with `reverts` naming this one. Members are in the order the reverts were made; `order` is the landing order."
          },
          "400": {
            "description": "{ error } — `key` is not a valid changeset key"
          },
          "404": {
            "description": "No such changeset, or a member the caller may not write"
          },
          "409": {
            "description": "{ error, conflicts? } — `changeset is {state}: nothing of it has landed` for an open, landing or abandoned changeset; `changeset {key} already exists`; `nothing of changeset {key} is landed: every member that landed was reverted` for a failed changeset the unwind fully put back; `{repo}/{change}: the repository no longer exists`. When a member cannot be reverted cleanly, `error` names the first in landing order — `{repo}/{change}: {ref} has changed since it landed at {paths}`, `{repo}/{change}: {ref} no longer exists`, `{repo}/{change}: revert/{key} already exists` — and `conflicts` lists every such member as { repo, change, why, changed?: [path], no_branch?, branch_exists? }. Nothing was written."
          }
        }
      }
    },
...
  }
}